quick fix
new ai and movement but fixing the build lfs issue with this commit and deleting the old branch
