+++ Assets/Prefabs/Crawlers/Crawler.prefab
@@ -270,8 +270,8 @@ MonoBehaviour:
   meshRenderer: {fileID: 0}
   target: {fileID: 0}
   crawlerSpawner: {fileID: 0}
-  attackDamage: 1
-  attackDistacne: 5
+  attackDamage: 2
+  attackRange: 5
   speed: 170
   randomScale: 0.2
   DeathBlood: {fileID: 7485218815324124808}
@@ -280,7 +280,6 @@ MonoBehaviour:
   dead: 0
   deathNoise: {fileID: 418313318034436250}
   immune: 0
-  randomLocationRadius: 3
   crawlerScale: 0.65
   cashWorth: 5
   expWorth: 3
@@ -294,6 +293,7 @@ MonoBehaviour:
   damageNumbersOn: 1
   _targetHealth: {fileID: 0}
   triggeredAttack: 0
+  _crawlerBehavior: {fileID: 0}
   spawnLocation: {x: 0, y: 0, z: 0}
 --- !u!54 &4026478879676073532
 Rigidbody:
@@ -304,7 +304,7 @@ Rigidbody:
   m_GameObject: {fileID: 3941810313144788156}
   serializedVersion: 4
   m_Mass: 2
-  m_Drag: 5.5
+  m_Drag: 5
   m_AngularDrag: 1
   m_CenterOfMass: {x: 0, y: 0, z: 0}
   m_InertiaTensor: {x: 1, y: 1, z: 1}
@@ -381,13 +381,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cad473b26ffd16d499fb9d7c6d864494, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  movementTarget: {fileID: 0}
   destination: {x: 0, y: 0, z: 0}
   speedFinal: 0
   steerSpeed: 60
   lookSpeed: 200
   stoppingDistance: 4
   canMove: 1
-  canSwarm: 1
   direction: {x: 0, y: 0, z: 0}
   rayAmount: 9
   rayDistance: 3
@@ -397,8 +397,6 @@ MonoBehaviour:
   distanceToTarget: 0
   tracking: 0
   separationWeight: 1
-  alignmentWeight: 1
-  cohesionWeight: 1
   swarmRadius: 20
   desiredSeparation: 5
   icedEffect: {fileID: 0}
@@ -406,6 +404,7 @@ MonoBehaviour:
   slowedAmount: 0.5
   m_crawler: {fileID: 0}
   obstacleAvoidanceWeight: 1
+  wanderRadius: 10
 --- !u!136 &8265754060228884709
 CapsuleCollider:
   m_ObjectHideFlags: 0
@@ -460,8 +459,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   baseAggression: 1
   healthAggressionModifier: 0.3
-  preferredEngageDistance: 10
-  maxEngageDistance: 20
+  preferredEngageDistance: 2
+  maxEngageDistance: 3
   fleeDistance: 15
   fleeHealthThreshold: 0.3
   cautionHealthThreshold: 0.5
