+++ Assets/Prefabs/Crawlers/CrawlerHolo.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 492426d676d87c74c82d4031a9526265
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
