+++ Assets/Prefabs/Props/Desert/Dead Tree.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4f72e1aca7fcc064095fd0cdb243dc32
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
