+++ Assets/Scripts/Crawlers/crawler-behaviour.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4ff8810e624eb1498ff84a5796aca3c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
