+++ Assets/Scripts/Props/Crate.cs
@@ -40,4 +40,5 @@ public class Crate : Prop
         yield return new WaitForSeconds(delay);
         gameObject.SetActive(false);
     }
+
 }
